pyproject.toml
[project]
name = "evm-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
description = "EVM blockchain data sync pipeline using AWS CDK"

dependencies = [
  "aws-cdk-lib>=2.100.0",
  "constructs>=10.0.0",
  "boto3>=1.28.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
]
lambda = [
  "requests>=2.31.0",
  "pandas>=2.0.0",
  "pyarrow>=15.0.0",
  "deltalake>=0.17.0",
  "web3>=6.0.0",
  "eth-utils>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["stacks*", "lambdas*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests as integration tests (require LocalStack)",
]
